refactor: establish HTTP API boundary
