+++ CHANGELOG.md
@@ -7,6 +7,7 @@ All notable changes to the Vibe Blog project will be documented in this file.
 ## 2026-07-25
 
 ### Changed
+- ♻️ **HTTP API 能力边界** — 将 Flask 应用工厂与 13 个 Blueprint 路由归入 `api/`，建立 schemas/errors 边界，修正静态资源路径并保留旧 `routes.*` patch 与导入兼容。
 - ♻️ **文档与发布能力边界** — 将解析、知识与 Book 能力归入 `services/documents/`，OSS、XHS 与发布工作流归入 `services/publishing/`，修正资源路径并保留旧导入兼容。
 - ♻️ **媒体服务能力边界** — 将图片、图片风格、视频、视频序列和 Sora 实现归入 `services/media/`，修正资源与 OSS 依赖路径，并以模块别名保留旧入口兼容。
 - ♻️ **调度与任务存储边界** — 将队列、cron 和调度编排迁入 `services/scheduling/`，任务模型及 SQLite 持久化分别归入 `models/` 与 `repositories/tasks/`，并保留旧导入兼容。
