+++ backend/api/errors.py
@@ -1,0 +1,1 @@
+"""Shared API error boundary for incremental route migration."""
