+++ backend/api/schemas/__init__.py
@@ -1,0 +1,1 @@
+"""Request and response schema boundary for incremental API migration."""
